+++ Rand_Templates/RandFixFloat.py
