+++ WDI/Zestaw_1/Zadanie_20/test_smallRand.py
@@ -1,6 +1,6 @@
 # Szymon Rusiecki
 import pytest
-from ....randTemplates.RFloat import RandFixFloat
+from ....Rand_Templates.RandFixFloat import RandFixFloat
 from ....readstdout import checkstdout
 from .prog import f as user_sol
 from .sol import f as corr_sol
@@ -15,7 +15,7 @@ SMALL_RAND_TESTS = [
 
 
 @pytest.mark.order(2)
-@pytest.mark.dependency(name="testSmallRand_s1t20", depends=["testBasic_s1t20"], scope="session")
+@pytest.mark.dependency(name="test_small_rand_s1t20", depends=["test_basic_s1t20"], scope="session")
 @pytest.mark.parametrize("data", SMALL_RAND_TESTS)
 def test_small_rand(data):
     assert checkstdout(user_sol, corr_sol, data, float_type=True)
